feat: new route and departure view structure for pattern selector and tabs
