+++ app/component/RouteStopListContainer.js
@@ -107,9 +107,7 @@ class RouteStopListContainer extends React.PureComponent {
 
   render() {
     return (
-      <div
-        className={cx('route-stop-list momentum-scroll', this.props.className)}
-      >
+      <div className={cx('route-stop-list', this.props.className)}>
         {this.getStops()}
       </div>
     );
