+++ app/component/TripStopListContainer.js
@@ -151,9 +151,7 @@ class TripStopListContainer extends React.PureComponent {
 
   render() {
     return (
-      <div
-        className={cx('route-stop-list momentum-scroll', this.props.className)}
-      >
+      <div className={cx('route-stop-list', this.props.className)}>
         {this.getStops()}
       </div>
     );
