+++ app/component/route.scss
@@ -57,20 +57,25 @@ $margin-left-right: 1.25em;
 }
 
 .route-page-content {
+  overflow-x: hidden;
+  overflow-y: auto;
   padding: 1.5em $margin-left-right 0;
   display: flex;
   position: relative;
   flex-direction: column;
-  max-height: 100%;
-  flex-grow: 1;
-  flex-basis: 100%;
+  padding-bottom: 1.5em;
+  flex: 1 0;
+  flex-basis: 0px;
+  max-height: 100vh;
   background: $white;
   transition: flex-grow 500ms cubic-bezier(0.215, 0.61, 0.355, 1),
     flex-basis 500ms cubic-bezier(0.215, 0.61, 0.355, 1);
 
   &.bp-large {
     top: 0;
-    padding: 0 $margin-left-right-large;
+    padding-top: 0;
+    padding-left: 3.5em;
+    padding-right: 2.5em;
   }
   &:empty {
     flex-basis: 0;
@@ -86,21 +91,14 @@ $margin-left-right: 1.25em;
 // stop list
 .route-stop-list {
   @include font-narrow-medium;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 1.5em;
-  overflow-x: hidden;
-  overflow-y: auto;
-  flex: 1 0;
-  flex-basis: 0px;
   background-color: $white;
   .route-stop:last-child .route-stop-row_content-container {
     border-bottom: none;
   }
 }
 
-.mobile .route-stop-list {
+.mobile .route-page-content {
+  padding-top: 0;
   overflow-y: visible;
   flex-basis: auto;
 }
@@ -395,6 +393,10 @@ $margin-left-right: 1.25em;
   @media print {
     display: block;
   }
+
+  .route-page-control-panel-container {
+    padding: 10px 1.25em 0;
+  }
 }
 
 .route-schedule-list-wrapper,
@@ -662,8 +664,22 @@ $margin-left-right: 1.25em;
   background-color: white;
   padding: 5px 20px 0 20px;
   &.bp-large {
-    padding: 0 $margin-left-right-large 1.5em;
+    padding: 0 $margin-left-right-large 0;
     box-shadow: none;
+  }
+}
+
+.route-page-control-panel-container {
+  padding-bottom: 1.5em;
+  &.hairiot, &.aikataulu {
+    padding-left: 1.5em;
+    padding-right: 1.5em;
+  }
+  &.bp-large {
+    &.hairiot, &.aikataulu {
+      padding-left: 3.5em;
+      padding-right: 3.5em;
+    }
   }
 }
